requirements.txt
accelerate==0.27.2
datasets==2.14.5
evaluate==0.4.1
matplotlib==3.8.0
numpy==1.26.4
pandas==2.2.1
Pillow==10.2.0
torch==2.2.1
torchvision==0.17.1
tqdm==4.66.1
transformers==4.38.1
tensorboard==2.15.0
